Code refactor - Update HtmlField.js   1. move draft editor component to editor/draft-editor   2. move editor buttons to editor/editor-buttons - move all the folders and files related to draft editor to editor/ - mv CONSTANTS.js to editor/entities.js
