+++ fields/types/html/editor/entity-decorator.js
