+++ fields/types/html/editor/info-box/info-box-bt.js
