+++ fields/types/html/editor/info-box/info-box-editing-block.js
