+++ fields/types/html/editor/modifiers/remove-block.js
