+++ fields/types/html/editor/quote/quote-decorator.js
